# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Notion KB Manager API
  description: |
    Production-ready API for managing knowledge base content with Notion integration.

    ## Features
    - 🔗 Link management and validation
    - 📄 Content parsing and extraction
    - 🤖 AI-powered content processing
    - 📝 Notion database integration
    - 💾 Backup and restore
    - ⚙️ Configuration management
    - 📊 Task monitoring and reporting

    ## Authentication
    Most endpoints require JWT authentication. Include your token in the `Authorization` header:
    ```
    Authorization: Bearer <your_token>
    ```

    To obtain a token, use the `/api/auth/login` endpoint with valid credentials.

    ## Rate Limiting
    API requests are rate-limited to prevent abuse:
    - Default: 1000 requests/hour, 100 requests/minute
    - Config endpoints: 100 requests/hour
    - Parsing: 10 requests/minute
    - AI processing: 20 requests/minute
    - Backups: 5 requests/hour

    Rate limit headers are included in responses:
    - `X-RateLimit-Limit`: Request limit
    - `X-RateLimit-Remaining`: Remaining requests
    - `X-RateLimit-Reset`: Reset timestamp

  version: 1.0.0
  contact:
    name: Notion KB Manager Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://staging-api.example.com/api
    description: Staging server
...
    description: Production server
...
security:
  - BearerAuth: []
paths:
...
  /tasks/pending:
    get:
      tags: [Tasks]
      summary: List pending tasks
      description: Get list of pending and queued tasks
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
        - name: task_type
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Pending tasks retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
  /tasks/pending/{task_id}:
    get:
      tags: [Tasks]
      summary: Get pending task details
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Task details retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Task'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Tasks]
      summary: Edit pending task
      description: Update task name or configuration
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                config:
                  type: object
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Tasks]
      summary: Delete pending task
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Task deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /tasks/history:
    get:
      tags: [Tasks]
      summary: List historical tasks
      description: Get list of completed and failed tasks
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [completed, failed]
      responses:
        '200':
          description: Historical tasks retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
  /tasks/history/{task_id}/rerun:
    post:
      tags: [Tasks]
      summary: Clone and rerun task
      description: Clone completed/failed task for re-execution
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Optional new task name
      responses:
        '200':
          description: Task cloned successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      task_id:
                        type: integer
                      status:
                        type: string
        '404':
          $ref: '#/components/responses/NotFound'
  /tasks/history/{task_id}/export:
    post:
      tags: [Tasks]
      summary: Export task report
      description: Generate Excel or JSON report for task execution
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                format:
                  type: string
                  enum: [excel, json]
                  default: excel
      responses:
        '200':
          description: Report generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      report_path:
                        type: string
                      download_url:
                        type: string
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== Monitoring ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/login
  parameters:
...
    TaskId:
      name: task_id
      in: path
      required: true
      schema:
        type: integer
      description: Task ID
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          type: object
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: string
    PaginatedResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            type: object
        pagination:
          type: object
          properties:
            page:
              type: integer
            per_page:
              type: integer
            total:
              type: integer
            pages:
              type: integer
...
    Task:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        task_type:
          type: string
        status:
          type: string
          enum: [pending, queued, running, completed, failed]
        created_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        error_message:
          type: string
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: REQ_002
              message: Missing required fields
              details: Field 'url' is required
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: RES_001
              message: Resource not found
              details: Link with ID 123 not found
...